move foreword to announcement section and credits bottom of home
